we can do
Yes, you can use this code in C++ with the "bits/stdc++.h" header file. This header file includes all the standard headers in C++ and is commonly used in competitive programming. However, it is not recommended to use this header file in production code or large projects, as it can slow down the compilation process.

The code itself calculates the maximum and minimum values of two integers, "a" and "b", using the formula:

Maximum = (a + b + abs(a - b)) / 2
Minimum = (a + b - abs(a - b)) / 2
The "abs" function calculates the absolute value of the difference between "a" and "b". The calculated values are then printed using the printf function.